SRAM, sim_top: Copy into main src directory
Copy the HASTI SRAM (memory) and the top-level module
(vscale_sim_top) into the src/main/verilog/ directory to make
it easier for RTLCheck to work with them.

Also rename vscale_sim_top.v to vscale_sim_top_unmod.v, as it is the
file to which assertions and assumptions are automatically added by
RTLCheck.