Multicore: Make V-scale Multicore, fix memory bug, add halt instr
Make the V-scale processor multicore (i.e. Multi-V-scale) by
instantiating 4 copies of the core and connecting them to memory via
an arbiter (vscale_arbiter). This includes changes to the SRAM
(vscale_dp_hasti_sram) and top-level module (vscale_sim_top).

This commit includes the following noteworthy changes in addition to
making V-scale multicore:

-A fix of the bug in the V-scale memory discovered by RTLCheck, where
if two stores are sent to memory in successive cycles, then the first
of the two stores is dropped by the memory.
-A halt instruction and its implementation, so cores can be halted.